+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [0.0.3] - 2023-10-19
+- Upgrade python-lib to v2.10.1
+- Upgrade mongo-libs to v4.2.7
+- Replaced the arg_parser code with gen_class.ArgParser code.
+- Updated to work in Python 3 too
+
+### Changed
+- single_db: Update arguments to mongo_libs.create_cmd call.
+- Converted imports to use Python 2.7 or Python 3.
+- Multiple functions: Replaced the arg_parser code with gen_class.ArgParser code.
+- main, single_db: Removed gen_libs.get_inst call.
+- Documentation updates.
+
+
 ## [0.0.2] - 2021-12-22
 - Upgrade mongo-libs to v4.2.1
 - Upgrade python-lib to v2.9.2